requirements.txt
fastapi
uvicorn
scikit-learn==1.6.1
joblib
numpy
python-whois
dnspython
tldextract